Add Remove student function and major refractor
Add RemoveStudentCommand and RemoveStudentCommandParser to facilitate
removing student.

How it works: The user's input to remove a student will be parsed by the
RemoveStudentCommandParser to obtain the student's name and its class.
The Parser returns a RemoveStudentCommand that removes the student from the
class upon execution.

Refractored Person.java to Student.java, some functions related to
Person also renamed to Student. Will continue to rename as time goes on
as there is too much to remove at one go without risking breakage of
code.

TempClass.java is put in place as a stub until code is merged with
actual implementation of Class object.